Update Process_MSG.ino
changed to int64_t so that uint32_t can be passed to the controller. rollover will be necessary every 70mins

+++ Process_MSG.ino
@@ -6,13 +6,13 @@ int secondInt[116] = { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
 //int MessagePOS;
 //byte Message[5];
 
-int32_t IntVals[8];
+int64_t IntVals[8];
 String StrVals[2];
 
 #include <FastCRC.h>
 FastCRC16 CRC16;
 
-int32_t ProcessNextInt() {
+int64_t ProcessNextInt() {
   byte ByteComb[5];
   int VQLp = 0;
